More comments and splitting up of files
