+++ .gitignore
@@ -42,6 +42,8 @@ Makefile
 *.swo
 *.swp
 src/compile_commands.json
+src/external/external
+.cache/
 
 # Emacs
 \#*#
