+++ src/software/ai/hl/stp/tactic/tactic.h
@@ -1,12 +1,10 @@
 #pragma once
 
 #include <boost/coroutine2/all.hpp>  // TODO (#1888): remove this dependency
-#include <functional>
-#include <include/boost/sml.hpp>
-#include <optional>  // TODO (#1888): remove this dependency
-#include <queue>
+#include <optional>                  // TODO (#1888): remove this dependency
 
 #include "software/ai/hl/stp/action/action.h"  // TODO (#1888): remove this dependency
+#include "software/ai/hl/stp/tactic/tactic_fsm.h"
 #include "software/ai/hl/stp/tactic/tactic_visitor.h"
 #include "software/ai/hl/stp/tactic/transition_conditions.h"
 #include "software/ai/intent/intent.h"
@@ -16,44 +14,6 @@
 // We typedef the coroutine return type to make it shorter, more descriptive,
 // and easier to work with
 typedef boost::coroutines2::coroutine<std::shared_ptr<Action>> ActionCoroutine;
-
-// alias for FSMs that have at least 2 levels of hierarchy
-template <class FSM>
-using HFSM = boost::sml::sm<FSM, boost::sml::process_queue<std::queue>>;
-
-// alias for FSMs that have no hierarchy
-template <class FSM>
-using BaseFSM = boost::sml::sm<FSM>;
-
-using SetIntentFunction = std::function<void(std::unique_ptr<Intent>)>;
-
-// The tactic update struct is used to update tactics and set the new intent
-struct TacticUpdate
-{
-    TacticUpdate(const Robot &robot, const World &world,
-                 const SetIntentFunction set_intent_fun)
-        : robot(robot), world(world), set_intent(set_intent_fun)
-    {
-    }
-    // updated robot that tactic is assigned to
-    Robot robot;
-    // updated world
-    World world;
-    // callback to return the next intent
-    SetIntentFunction set_intent;
-};
-
-#define UPDATE_STRUCT_WITH_CONTROL_PARAMS_AND_COMMON                                     \
-    struct Update                                                                        \
-    {                                                                                    \
-        Update(const ControlParams &control_params, const TacticUpdate &common)          \
-            : control_params(control_params), common(common)                             \
-        {                                                                                \
-        }                                                                                \
-        ControlParams control_params;                                                    \
-        TacticUpdate common;                                                             \
-    };
-
 
 /**
  * In the STP framework, a Tactic represents a role or objective for a single robot.
